rename : course 관련 파일 패키지 생성해서 정리
