+++ src/main/java/com/innocamp/dduha/model/course/CourseDetailSpot.java
@@ -1,6 +1,6 @@
-package com.innocamp.dduha.model;
+package com.innocamp.dduha.model.course;
 
-import com.innocamp.dduha.model.accommodation.Accommodation;
+import com.innocamp.dduha.model.touristspot.TouristSpot;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -13,17 +13,20 @@ import javax.persistence.*;
 @NoArgsConstructor
 @Getter
 @Builder
-public class CourseDetailAcc {
+public class CourseDetailSpot {
 
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
     private Long id;
 
+    @Column(nullable = false)
+    private int detailOrder;
+
     @JoinColumn(name = "course_id", nullable = false)
     @ManyToOne(fetch = FetchType.LAZY)
     private Course course;
 
-    @JoinColumn(name = "acc_id", nullable = false)
+    @JoinColumn(name = "spot_id", nullable = false)
     @ManyToOne(fetch = FetchType.LAZY)
-    private Accommodation accommodation;
+    private TouristSpot touristSpot;
 }
