+++ src/main/java/com/innocamp/dduha/repository/coursedetail/CourseDetailAccReposiotry.java
@@ -1,7 +1,7 @@
 package com.innocamp.dduha.repository.coursedetail;
 
-import com.innocamp.dduha.model.Course;
-import com.innocamp.dduha.model.CourseDetailAcc;
+import com.innocamp.dduha.model.course.Course;
+import com.innocamp.dduha.model.course.CourseDetailAcc;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CourseDetailAccReposiotry extends JpaRepository<CourseDetailAcc, Long> {
