+++ src/main/java/com/innocamp/dduha/service/TripService.java
@@ -11,6 +11,10 @@ import com.innocamp.dduha.dto.response.TripResponseDto;
 import com.innocamp.dduha.jwt.TokenProvider;
 import com.innocamp.dduha.model.*;
 import com.innocamp.dduha.model.bookmark.TripBookmark;
+import com.innocamp.dduha.model.course.Course;
+import com.innocamp.dduha.model.course.CourseDetailAcc;
+import com.innocamp.dduha.model.course.CourseDetailRest;
+import com.innocamp.dduha.model.course.CourseDetailSpot;
 import com.innocamp.dduha.model.restaurant.Restaurant;
 import com.innocamp.dduha.model.touristspot.TouristSpot;
 import com.innocamp.dduha.repository.accommodation.AccommodationRepository;
